requirements.txt
astroid==3.2.2
blinker==1.8.2
cachelib==0.9.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
dill==0.3.8
Flask==2.2.0
Flask-Caching==2.3.0
flask-ldap3-login==0.9.18
Flask-Limiter==3.7.0
Flask-WTF==1.2.1
idna==3.7
importlib_resources==6.4.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
ldap3==2.9.1
limits==3.12.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
pip==24.0
platformdirs==4.2.2
pyasn1==0.6.0
Pygments==2.18.0
pylint==3.2.3
PyYAML==6.0.1
redis==5.0.6
requests==2.32.3
rich==13.7.1
setuptools==65.5.0
tomlkit==0.12.5
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==2.2.0
wrapt==1.16.0
WTForms==3.1.2
